requirements.txt
Django==5.2.3
djangorestframework==3.15.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.8.0
bcrypt==4.0.1
